Rename `build.py` to `build_go_pkg.py`
This isn't the ideal final destination, but it can't go into `util_rules/go_pkg.py` because of a circular import.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
